MAJOR: refactor function locations and correct `stac_api` parameter to `stac_api_name` where applicable.
